Simplified seed db script. Trying to correct issues that appear
in production on the linux server. Have yet to have the seed
with credentials fail in a dev environment. I feel this is a
TOCTOU issue. Therefore I added timeouts for any seeds that rely
on other tables to ensure that sequelize has a chance to catch up

Have tried the new script multiple times on a dev rocky linux
server and have not had any issues with the seeds failing.