Fixed up failing tests. Corrected issue with unique validation
faliure on the paranoid tables. When making an employee an admin
or supervisor after they have lost supervisor status was causing
an error.

The following tables are no longer soft deleted:
  - Supervisors
  - Login Credentials
  - Create Credentials Invites

They have unique constraints, and if they are not really deleted
readding or editing some models will cause validation issues
due to unique constraints. These tables are not really that
important. If a row is deleted, it can be recreated without
extreme data loss. The employees, and their CallOuts remain
in a paranoid state, making them harder to delete, their only
unique constraints are in their ids and should cause no issues
as ids are UUIDs