Created production scripts
Created an install-as-service script that installs dependencies,
builds the JS, generates the TLS certificates if needed,
and creates a systemd service for Rocky Linux.

Created an update script that handles fetching updates from
GitHub, rebuilding the project, and restarting the system service.
SELinux policies are created so systemd will run the service.

Created a clean script that only removes build artifacts ie, the
.next folder and the dist folder.

Created an uninstall script, that completely removes the project
except for the project's source files and removes the service
and SELinux policies created.